Add `read/write_array::<Length>()` to save one byte
