+++ Cargo.toml
@@ -6,5 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+generic-array = "0.14.6"
 leb128 = "0.2.5"
 thiserror = "1.0.32"
